+++ apps/questdk/src/index.ts
@@ -16,6 +16,7 @@ export type {
   OptionsActionParams,
   DelegateActionParams,
   VoteActionParams,
+  CreateActionParams,
   ActionParams,
   IActionPlugin,
 } from './actions'